feat(): extract signing process
